Fix the bug raised at #3451
"TypeError: can't multiply sequence by non-int of type 'float'";

It happened after my change to that expression inside # #3318.

+++ instapy/util.py
@@ -926,7 +926,7 @@ def highlight_print(username=None, message=None, priority=None, level=None, logg
 
 
     if show_logs == True:
-        print("\n{}".format(upper_char * ceil(output_len/len(upper_char))))
+        print("\n{}".format(upper_char * int(ceil(output_len/len(upper_char)))))
 
     if level == "info":
         logger.info(message)
